updated card table formats
updated card table formats with marketplace functionality (tested in production)
