+++ html/artwork/allseries/index.php
@@ -19,7 +19,7 @@ require_once('libraries/stripe/init.php');
 \Stripe\Stripe::setApiKey($apikey['stripe']['secret']);
 
 /* PAGE VARIABLES */
-$currentpage = 'artwork/';
+$currentpage = 'artwork/allseries/';
 
 
 
@@ -39,12 +39,13 @@ ob_end_flush();
 
 
 /* HEADER */ require('layout/header2.php');
-/* HEADER */ require('layout/header1.php');
+/* HEADER */ require('layout/header1.php'); 
 
 print'
 
-<script language="javascript" type="text/javascript" src="table.js"></script>
-<link rel="stylesheet" type="text/css" href="sorting.css">
+<script language="javascript" type="text/javascript" src="https://helmarbrewing.com/js/jquery-1.12.4.js"></script>
+<script language="javascript" type="text/javascript" src="https://helmarbrewing.com/js/jquery.dataTables.min.js"></script>
+<link rel="stylesheet" type="text/css" href="https://helmarbrewing.com/js/jquery.dataTables.min.css">
 
 
 <div class="artwork">
@@ -59,14 +60,13 @@ print'
 
  /* setup code if 1) user logged in with no subscription, 2) user logged in with subscription, 3) user not logged in */
 
- if(isset($user)){
-    if( $user->login() === 1 || $user->login() === 2 ){
-		/* do this code if user is logged in */
+
 
 		print'
 			<div class="series_desc">Below is the Helmar Brewing checklist for all of our cards.
 			You can use filters to display the cards you want and you can even download the checklist.
 			To get more information regarding each series, <a href="'.$protocol.$site.'/artwork/">please visit the individual series pages</a>.</div>';
+
 
 		// load active auctions in array
 		$R_cards2 = $db_main->query("
@@ -83,15 +83,12 @@ print'
 
 		}
 
+
 		print '<div class="series_desc"><p><a href="/artwork/csvall/"><i class="fa fa-download"></i> Download Complete Card List</a></p></div>';
 
 
 	$R_cards = $db_main->query("
-	SELECT cardList.*, userCardChecklist.quantity, userCardChecklist.wishlistQuantity, series_info.series_name FROM cardList
-	LEFT JOIN userCardChecklist
-		ON cardList.series = userCardChecklist.series
-		AND cardList.cardnum = userCardChecklist.cardnum
-		AND userCardChecklist.userid = '".$user->id."'
+	SELECT cardList.*, series_info.series_name FROM cardList
 	LEFT JOIN series_info
 			ON cardList.series = series_info.series_tag"
 );
@@ -101,21 +98,17 @@ print'
 				// http://www.javascripttoolbox.com/lib/table/examples.php
 				//
 
+
             print'
-						<table id="t1" class="tables table-autosort table-autofilter table-stripeclass:alternate table-page-number:t1page table-page-count:t1pages table-filtered-rowcount:t1filtercount table-rowcount:t1allcount">
+						<table id="example" class="display compact">
                       <thead>
                         <tr>
-													<th class="table-filterable table-sortable:ignorecase">Series</th>
-                          <th class="table-sortable:numeric">Card Number</th>
-                          <th class="table-filterable table-sortable:ignorecase">Player</th>
-                          <th class="table-filterable table-sortable:ignorecase">Stance / Position</th>
-                          <th class="table-filterable table-sortable:ignorecase">Team</th>
-                          <th class="table-sortable:date">Last Sold Date</th>
-                          <th class="table-sortable:currency">Max Sell Price</th>
-                          <th>Pictures</th>
-            			        <th>Personal Checklist</th>
-													<th>Personal Wishlist</th>
-													<th class="table-sortable:numeric">Active Auction</th>
+													<th>Series</th>
+                          <th>Card #</th>
+                          <th>Player</th>
+                          <th>Stance / Position</th>
+                          <th>Team</th>
+													<th>Pictures</th>
                         </tr>
                       </thead>
                       <tbody>
@@ -132,81 +125,51 @@ print'
                             <td>'.$card->description.'</td>
                             <td>'.$card->team.'</td>
                     ';
-                    if($card->averagesold == 0){
-                        print'
-                            <td></td>
-                            <td></td>
-                        ';
-                    }else{
-                        print'
-                            <td>'.$card->lastsold.'</td>
-                            <td>'.$card->maxSold.'</td>
-                        ';
-                    }
 
-                    // need to add ************ for one row, check if picture exists
-                    print'
-                            <td align="center" class="picCol">
-                    ';
+										// need to add ************ for one row, check if picture exists
+										print'
+														<td align="center" class="picCol">
+										';
 
 
-                    // define the pictures
-                    $frontpic = '/images/cardPics/'.$card->series.'_'.$card->cardnum.'_Front.jpg';
-                    $frontthumb = '/images/cardPics/thumb/'.$card->series.'_'.$card->cardnum.'_Front.jpg';
-                    $backpic  = '/images/cardPics/'.$card->series.'_'.$card->cardnum.'_Back.jpg';
-                    $backthumb  = '/images/cardPics/thumb/'.$card->series.'_'.$card->cardnum.'_Back.jpg';
+										// define the pictures
+										$frontpic = '/images/cardPics/'.$card->series.'_'.$card->cardnum.'_Front.jpg';
+										$frontthumb = '/images/cardPics/thumb/'.$card->series.'_'.$card->cardnum.'_Front.jpg';
+										$backpic  = '/images/cardPics/'.$card->series.'_'.$card->cardnum.'_Back.jpg';
+										$backthumb  = '/images/cardPics/thumb/'.$card->series.'_'.$card->cardnum.'_Back.jpg';
 										$frontlarge = '/images/cardPics/large/'.$card->series.'_'.$card->cardnum.'_Front.jpg';
 										$backlarge  = '/images/cardPics/large/'.$card->series.'_'.$card->cardnum.'_Back.jpg';
 
-                    //check if either pic exists
-                    if( file_exists($_SERVER['DOCUMENT_ROOT'].$frontlarge) || file_exists($_SERVER['DOCUMENT_ROOT'].$backlarge) ){
+										//check if either pic exists
+										if( file_exists($_SERVER['DOCUMENT_ROOT'].$frontlarge) || file_exists($_SERVER['DOCUMENT_ROOT'].$backlarge) ){
 
-                        // print the front pic if exists
-                        if(file_exists($_SERVER['DOCUMENT_ROOT'].$frontlarge)){
-                            print'
-                                <a href="'.$protocol.$site.'/'.$frontlarge.'" data-lightbox="'.$card->series.'_'.$card->cardnum.'" ><img src="'.$protocol.$site.$frontthumb.'"></a>
-                            ';
-                        }
+												// print the front pic if exists
+												if(file_exists($_SERVER['DOCUMENT_ROOT'].$frontlarge)){
+														print'
+																<a href="'.$protocol.$site.'/'.$frontlarge.'" data-lightbox="'.$card->series.'_'.$card->cardnum.'" ><img src="'.$protocol.$site.$frontthumb.'"></a>
+														';
+												}
 
-                        // insert space
-                        if( file_exists($_SERVER['DOCUMENT_ROOT'].$frontlarge) && file_exists($_SERVER['DOCUMENT_ROOT'].$backlarge) ){
-                            print'&nbsp;&nbsp;';
-                        }
+												// insert space
+												if( file_exists($_SERVER['DOCUMENT_ROOT'].$frontlarge) && file_exists($_SERVER['DOCUMENT_ROOT'].$backlarge) ){
+														print'&nbsp;&nbsp;';
+												}
 
-                        // print the back pic if exists
-                        if(file_exists($_SERVER['DOCUMENT_ROOT'].$backlarge)){
-                            print'
-                                <a href="'.$protocol.$site.'/'.$backlarge.'" data-lightbox="'.$card->series.'_'.$card->cardnum.'" ><img src="'.$protocol.$site.$backthumb.'"></a>
-                            ';
+												// print the back pic if exists
+												if(file_exists($_SERVER['DOCUMENT_ROOT'].$backlarge)){
+														print'
+																<a href="'.$protocol.$site.'/'.$backlarge.'" data-lightbox="'.$card->series.'_'.$card->cardnum.'" ><img src="'.$protocol.$site.$backthumb.'"></a>
+														';
 
-                        }
+												}
 
-                    // neither pic exists print message instead
-                    }else{
-                        print'
-                                <i>no picture</i>
-                        ';
-                    }
-            		/* end card pic */
-                    print'</td>';
-
-
-            		/* add CHECKLIST icon */
-					if($card->quantity > 0){
-						print '<td><a href="javascript:;"><i class="fa fa-check-square-o" onclick="checklist(\''.$card->series.'\',\''.$card->cardnum.'\')" id="'.$card->series.'-'.$card->cardnum.'"></i></a></td>';
-					} else{
-						print '<td><a href="javascript:;"><i class="fa fa-square-o" onclick="checklist(\''.$card->series.'\',\''.$card->cardnum.'\')" id="'.$card->series.'-'.$card->cardnum.'"></i></a></td>';
-					}
-
-					/* add WISHLIST icon */
-		if($card->wishlistQuantity > 0){
-			print '<td><a href="javascript:;"><i class="fa fa-check-square-o" onclick="wishlist(\''.$card->series.'\',\''.$card->cardnum.'\')" id="WISH'.$card->series.'-'.$card->cardnum.'"></i></a></td>';
-		} else{
-			print '<td><a href="javascript:;"><i class="fa fa-square-o" onclick="wishlist(\''.$card->series.'\',\''.$card->cardnum.'\')" id="WISH'.$card->series.'-'.$card->cardnum.'"></i></a></td>';
-		}
-
-		// add ebay auction if active
-			print '<td><a href="http://www.ebay.com/itm/'. $wishlist[$card->series.'-'.$card->cardnum]['auctionID'].'/" target="_blank">'.$wishlist[$card->series.'-'.$card->cardnum]['auctionID'].'</a></td>';
+										// neither pic exists print message instead
+										}else{
+												print'
+																<i>no picture</i>
+												';
+										}
+								/* end card pic */
 
 
             		/* end row */
@@ -237,23 +200,11 @@ print'
 
 
 
-		/* END code if user is logged in, but not paid subscription */
-    }else{
-		/* do this if user is not logged in */
-		print'
-			<div class="series_desc">To view the Helmar Brewing checklist, <a href="'.$protocol.$site.'/account/register/">please sign up for a free account or log in</a>!</div>';
-		print '<a href="'.$protocol.$site.'/account/register/"><img class="centered" src="'.$protocol.$site.'/img/checklist-sample.jpg"></a>';
-	}
-}else{
-		/* do this if user is not logged in */
-		print'
-			<div class="series_desc">To view the Helmar Brewing checklist, <a href="'.$protocol.$site.'/account/register/">please sign up for a free account or log in</a>!</div>';
-		print '<a href="'.$protocol.$site.'/account/register/"><img class="centered" src="'.$protocol.$site.'/img/checklist-sample.jpg"></a>';
-}
-
 print'
 	</div>
 ';
+
+
 
 
 /* FOOTER */ require('layout/footer1.php');
@@ -261,6 +212,23 @@ print'
 $db_auth->close();
 $db_main->close();
 ?>
+
+<script>
+$(document).ready(function() {
+    $('#example').DataTable({
+  "columns": [
+    { "width": "25%" },
+    { "width": "7%" },
+    { "width": "23%" },
+    { "width": "20%" },
+    { "width": "20%" },
+		{ "width": "10%" }
+  ],
+  "pageLength": 50
+});
+} );
+</script>
+
 
 <script>
 function checklist(s, c){
